Comments - Defect Number: 2973; Defect RecordID: 2969; Server Address: services.logic2020.com; Server Port: 99; Database Name: Practice Management; Defect Summary: Report: Consulting Demand --> Add a link to the Opportunity Details page even for Projects; Defect Description: In the original specification, I requested that when a Project (either Active or Projected) is displayed under a particular Strawman resource on the Consulting Demand report, a hyperlink to the respective Project Details page should be displayed next to the Project. That was successfully implemented. Feedback from end-users, however, has returned. They would also like to have a hyperlink to the linked Opportunity for that Project as well so that a user may open the linked Opportunity Detail or the Project Detail from the Consulting Demand Report.
Therefore, when we display a Project name on the Consulting Demand report, we should display both the hyperlink to the Project Detail and a hyperlink to the Opportunity Detail that is linked to that project. The Project Hyperlink should always appear in the left-most position (i.e. add the Opportunity hyperlink to the right of the existing Project hyperlink separated by a hyphen).

e.g. Microsoft-Sample Project Name - P012345 - O701111