Comments - Defect Number: 2964; Defect RecordID: 2960; Server Address: services.logic2020.com; Server Port: 99; Database Name: Practice Management; Defect Summary: Project Detail : Add Description text box; Defect Description: Please add a Description text entry area for Projects as well that will inherit from/sync with the Description text entry area for an Opportunity if a linking relationship exists.
If the Project is created and then linked to an Oppty, prompt the user to choose one description over the other (helpful if you can show both to let them choose from).

Once the Project and Oppty are linked, changing the description in one should also change the other.

On the Consulting Demand report, show the description in a hover over the P##### linking.