Comments - Defect Number: 2980; Defect RecordID: 2976; Server Address: services.logic2020.com; Server Port: 99; Database Name: Practice Management; Defect Summary: New Enhancement - Improved Time Tracking; Defect Description: Madhan,
Attached is the draft specification for some significant changes to the way Logic20/20 wants to track time in Practice Management. Please review the Word document and reference the attached Excel workbook to provide additional detail. Ill step through each with you and your team tomorrow morning (Pacific time).

I imagine some impacts on our existing data model/schema. I will need some expert guidance from your team on the impacts of this specification and how best to architect the data model to provide the operational reporting flexibility we will need as a result of the system changes. We will want to be able to have a far more interactive reporting experience than what is presently afforded in PM to enable roll-up and drill-down capability leveraging this new time tracking design.

Additionally, I am attaching some sample UI changes we would like to discuss as these represent conceptual thoughts on the direction wed like to move for our Time Entry interface. These are not to be viewed as the only approach but should be seen as a sample for discussion. I actually think that combining elements of both is the right solution but lets discuss based on your reading of the specifications.

Our hope is that GGK will be able to run with this specification while Jason and I are on holiday from 12/22  1/2/2012 and return with some sample prototypes if not working code in a test environment by that first week of January. I can make myself available for a call during my holiday if the team gets stuck or has urgent questions that must be answered to avoid getting stuck.

Best,
Matt

Matthew Reilly, Senior Manager  Consulting Operations// Logic20/20 // MattR@logic2020.com
t.206-576-0491//c.425-503-5825// www.logic2020.com