Comments - Defect Number: 3053; Defect RecordID: 3049; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Change to project search for exact matches; Defect Description: If someone enters a project number into search that has an exact match, can the person be directed directly to the project detail? For example, if I enter "P123456" into the project search (case-insensative) and there is a project P123456 that exists, I would like to be directed to the project details page for that project instead of the search results.
Keyword search should still send people to the search results page.

Also, can the keyword search not be exact phrase? Currently if I enter "MCS FY12" in the search there are no results, though there is a project with both of those strings in the title, and that result should be returned. I am expecting that the search uses both of those strings with AND, and should not imply exact phrase.