Comments - Defect Number: 2981; Defect RecordID: 2977; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Termination process - when Person status changes to Terminated system should automatically adjust Persons end dates on all pro@; Defect Description: To terminate an employee in PM we must change the Person record status from Active to Terminated and enter the Termination Date. The system then checks all Project Records to determine if the person has project assignments that are beyond the termination date and checks whether they have time entered in Time Entry beyond the proposed termination date. The system provides an error dialog showing that the user has time entered beyond the termination date and the projects that still list the resource as assigned beyond the proposed termination date.
I want to change the process such that upon showing the above error the user is presented with the option to proceed with the Termination status change. If "Yes" then the system should automatically adjust the relevant end dates on the projects depicted for that individual resource as well as automatically delete the future dated time entry for the user. This will greatly reduce the manual entry process for terminating an employee. If "No" the system should cancel the operation and close the dialog box.