Comments - Defect Number: 3060; Defect RecordID: 3056; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Multiple results for user in By Time Period > By Resource report; Defect Description: Repro'd on multiple computers, report parameters are Payroll - Previous, showing persons with zero hours entered, filtering on Pay Type for W2-Hourly only. Kristin Cherry appears twice with exact same entries, but it's not clear why.
UPDATE: Matt suggested this might be due to Kristin having a new compensation record that takes effect on 7/1/12. If that is the reason, we do not want to have multiple entries for any user/employee. The "Pay Type" should report the most current pay type applicable to the range selected (in this case, 6/1-6/15), but since we are not doing any financial calculations in this report there should not be an issue with reporting raw number of hours entered in Time Entry.