Comments - Defect Number: 3063; Defect RecordID: 3059; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Configuration: Person - Termination should lockout user on Termination Date + 1 day; Defect Description: Currently, if a user terminates a person in the system, the user enters the Termination Date. Clicking Save will lock the person being terminated out of PM on the termination date itself. This is problematic as the person's employment technically ends at the end of that day, not the beginning. The person still needs access to the system to enable them to enter their time, finalize any last minute project/opportunity changes, etc.
Therefore, we'd like the system to lock the person out of the system on the next calendar day instead. (i.e. Termination Date + 1)

For example, if I terminate Employee Jane Doe with an Effective Date on July 2, 2012. Jane Doe will not be able to access PM on July 2nd. With this change, Jane Doe would retain access to PM through 7/2 and would not be able to access PM at 00:00:00 on July 3rd.