Comments - Defect Number: 3076; Defect RecordID: 3072; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Project Summary: Correct Excel Export to include all fields on UI and separate Revenue/Margin $ into separate cells; Defect Description: We need to improve the Project Summary page's Export. It does not include the Project's Start and End Dates in the Export and it does not separate Revenue and Margin $ into distinct cells which makes it very hard to work with as an Excel file without manual effort.
Need to ensure that the Revenue and Margin $ are properly formatted as $ values as well, not as "General" to enable use of calculations.
See the attached file for a view of what the current export layout is and what the desired layout is.

Please provide a level of effort assessment and when this change could be reflected in the UAT environment.