Comments - Defect Number: 3077; Defect RecordID: 3073; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Opportunity - Project Conversion: converting an oppty to a project is not porting the Business Unit; Defect Description: When a user selects the, "Convert This Opportunity to Project" button on the Opportunity Detail page, the system is failing to port the Group value (i.e. the Business Unit) to the Project Detail.
Steps to Reproduce:

Select an Opportunity and view its details on the Opportunity Detail page (use a Microsoft opportunity as they have multiple Group values to verify this defect easier)
Note the value in the "Group" field (this is the Business Unit) --> we should rename "Group" to be Business Unit as part of this defect fix
Select the "Convert This Opportunity to Project" button.
When the Project Details page appears, note the value in the "Business Unit" field. It has changed. In the case of a Microsoft opportunity it appears to always change to, "Azure (SQL)" presumably because this is the first Business Unit in the list.

Expected Outcome:
The Group value (whose field label should be renamed, Business Unit) should be the same on the Project Details page.