Comments - Defect Number: 2979; Defect RecordID: 2975; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Margin Test --> need to add an "Effective Date" field to the page; Defect Description: We need to add some temporal awareness to our Margin Test pages to take into consideration upcoming Compensation Record changes. This is especially poignant as we approach the performance evaluation timeframe each year when the vast majority of our consultants receive compensation adjustments. These are typically on January 1st and July 1st of each year but can happen at any point during the year. As the page is designed today, whenever a user enters values on the Margin Test page, the calculated values return based on the Compensation Record in effect that day. In fact, however, what is more appropriate is to enable a user to specify the effective date as we use this to project a project in the future. If we do not adjust to take into account a future dated Compensation Record than we are overstating our margins.
I am proposing that we add an "Effective Date" calendar field to each of the Margin Test pages. I believe it would be best to locate the Effective Date field just above the "Hourly Bill Rate (HBR)" field inside the "Revenue & Margin Report" section of each page.

The Effective Date field should default with today's date but allow a user to change the date as needed leveraging a calendar if needed or by manually entering a date. Once entered, the system should then calculate the values using the Compensation Record that is in effect for that date.

- Matt