Comments - Defect Number: 3082; Defect RecordID: 3078; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Change Employee Status - Compensation End Date is not deleted if user cancels Termination; Defect Description: When I canceled an individuals Termination Pending status via the Change Employee Status workflow, the compensation end date remains in the table.
Expected Result: Upon canceling a Termination, the person's previous compensation record end date should be deleted. (i.e. we need to restore an open compensation record if the termination is canceled)

Observed Result: The termination was canceled and the person's status returned to Active yet the compensation record still reflected the end date of the termination date.