Comments - Defect Number: 3086; Defect RecordID: 3082; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Margin - would like to expand visibility of Senior Managers to see Margin for all levels below Senior Manager; Defect Description: Currently, the rules governing who can see Margin % and Margin $ restricts a user to seeing Margins only for those two Seniority levels below them. We'd like to investigate how to implement a solution that would enable Senior Managers to see Margin $ and Margin % values for all people one level below them (e.g. Managers and below)
Can you research and confirm how the current algorithm works and recommend what would be needed to enable those with the Seniorty = Senior Manager to see one level below.

Senior Managers should NOT be able to see Margin $ or Margin % for any project that contains another Senior Manager (e.g. if two Sr. Managers are on the same project or when looking at another project containing a Senior Manager). If the user is the only Senior Manager on a project they should be able to see the Margin $ and Margin %.