Comments - Defect Number: 3087; Defect RecordID: 3083; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Add Project - what determines whether a user can change the Saleperson and/or Client Director fields when adding a project?; Defect Description: What determines whether a user can change the Saleperson and/or Client Director fields when adding a project?
It seems strange that if a user has the requisite permissions to Add a Project that they wouldn't have the permission to edit the Salesperson or Client Director fields when doing so. It also appears that some users can't enter the SOW Budget or select Project Manager(s) values.

Is it possible to make the following fields editable when Adding a Project and what are the impacts of doing so? (e.g. what is the business logic that would be changed and the impacts of doing so)
Account (already editable)
Salesperson
Client Director
Business Unit (already editable)
Buyer Name(already editable)
Project Owner(already editable)
Practice Area(already editable)
SOW Budget
Project Manager(s)
Project Notes(already editable)