Comments - Defect Number: 3096; Defect RecordID: 3092; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Project Details: Need a Capabilities field to track type of services performed by Project; Defect Description: We would like to track the capabilities we are executing on a project as part of the information we capture on the Project Details page.
The user should be able to select multiple capabilities for a single project.

See the attached file for a depiction of the Practice Area to Capability mapping. Each Capability is a child of its Practice Area parent.

I envision leveraging a multi-select dropdown similar to the "Project Managers" field. The content of the multi-select would then be a concatenation of the Practice Area name and the Capability Name sorted alphabetically (see the attachment for examples). I'm open to alternative user experiences, however, as I can't speak to how effective my recommendation is from a usability perspective and welcome your feedback.

As a second phase to this requirement, I also need a Configuration page for these capabilities that will enable an Administrator to add new capabilities or edit existing capabilities including name changes and associating each capability to one and only one parent Practice Area. Only Active Practice areas should be open for associating new capabilities.