Comments - Defect Number: 3079; Defect RecordID: 3075; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Reports: By Time Period report - add summary of Time-Off hours to the top of the report; Defect Description: We would like to see a total of the "Time-Off" hours displayed at the top of this report. One vision to satisfy this requirement is to add a "Total Time-Off" value next to the "Total Hours" value in the header. It would simply sum all of the hours in the "Time-Off" column.
Another possible vision is to provide summary values for all of the columns such that we would at a glance see the totals for:

Billable
Non-Billable
BD
Internal
Time-Off
Total

I'd be interested in seeing samples of each to decide which is most appropriate. I attempted to mock them up myself and welcome your thoughts/opinions.

Additionally, for the Time-Off value it would be nice to see the break-out summary of the Time-Off by Type - for example, if a user hovers over the "Total Time-Off" value (once implemented), the user would see that there were, for example, 120 hours of PTO, 8 hours of Unpaid, 16 hours of Sick-Leave, 24 hours of Jury Duty, and 0 hours of Bereavement.