Comments - Defect Number: 3092; Defect RecordID: 3088; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Opportunities - need to replace "A", "B", "C" etc. with actual sales stages; Defect Description: We are changing our overall sales discipline and need to update PM to correspond to the change in behavior we are driving within our sales team. Part of this change is the concept of a sales funnel which depicts the stages an opportunity traverses on its way to being Won or Lost.
We would like to replace the use of the current set of Priority options (PO, A, B, C, D...) and use the following stages instead:
Qualifying
Developing
Proposed
Negotiating
Closed
As part of this change we'd like to relabel the field from "Priority" to "Sales Stage". It should remain a required value.

I'd like assistance researching the impacts of adjusting our current opportunity labels to use the above names. I indicated the relative mapping to the existing Priority labels in parenthesis below next to each new value:

Qualifying (D)
Developing (C)
Proposed (B)
Negotiating (A)
Closed (PO)

I am not certain if application/business logic throughout the system relies on the hardcode values of "A", "B", "C", "D", and/or "PO" to drive other system behaviors. Or, ideally, executing this change is simply a Display Name update with little, if any, impact.

Also, is it possible to add additional Stages to expand beyond the 5 listed above?