Comments - Defect Number: 3078; Defect RecordID: 3074; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Time Entry: New Work Type needed exclusively for W2-Hourly employees to support new legislation; Defect Description: For W2 Hourly employees, the City of Seattle has passed new legislation which requires supporting Sick Leave even for hourly employees. As such we need to afford our W2 Hourly employees the ability to claim Sick leave in their Time Entry in Practice Management. We want to uniquely identify and keep it separate from PTO. Please see the attached mockup for the specification and impacts to both Time Entry and the Calendar's "Add Time Off" functionality.
Please contact Lisa Neel (lisaj@logic2020.com) with any questions of this specification during the week of August 27th through September 3rd as I will be on PTO. She does not have access to TTP so please e-mail her during this period directly.

The new Work Type code should be W9311 and should map to the Project Code, "PTO" which is P999905 and the Business Unit code of B9900 and Account Code of C2020.

- Matt