Comments - Defect Number: 3101; Defect RecordID: 3097; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Reports: Reports / Account  Business Development view's Expand All doesn't expand all; Defect Description: With reference to Defect #3054.
In the Reports / Account report view, click on the Business Development " tab for a given Account (e.g. Microsoft). If the user clicks Expand All, each person's results expands showing each day for which they entered time towards that Business Unit but the user should further expand to see the hours and notes breakdown for that day.

Expected Result: When a user clicks on the Business Development tab for a given Account, the individual business unit summary for the selected account should be displayed. And by clicking on the Expand All would then open each day revealing the time entry hours and the associated notes.