Comments - Defect Number: 3106; Defect RecordID: 3102; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Reports-->Human Capital: Terminations Report - numerator of the Attrition % calc is incorrectly including Contingent status co@; Defect Description: In the numerator and denominatorof the Attrition % calc, PM is incorrectly including those employees with a Status = Contingent in the count of terminations in the specified period.
Contingent status employees have not actually started with the company. Therefore, even though they may go to a terminated status, including them in the numerator/denominator will greatly exaggerate our actual Attrition %.


Action: The # of employee terminations during the specified time period should exclude those employees who were Contingent status but now have a status of Terminated. In other words, only employees who were Active and then went to Terminated status should be included in the numerator of the Attrition % calculation.

Example: In Production as of 11/15, the numerator is showing 79 terminations. However, when I remove those with Termination Reasons = "Contingent - ...", the terminations count reduces to 50. I would expect that 50 is used in the Attrition % calculation, not 79.