Comments - Defect Number: 3107; Defect RecordID: 3103; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Reports>Time>Create New "Payroll Distribution Report" SQL query Excel export; Defect Description: The Finance team has requested a data extract containing project information for every active employee in the company in a single Excel spreadsheet. The job is to be automatically run on the 3rd and the 18th of every month and emailed to Finance@logic2020.com. Please see mock-up attached.
The new report is based on the current report found by selecting Reports> Time > Reports > By Person > Detail tab > Export: Excel. Please use this report's output as a guide for creating a new job. Do NOT change the existing report as this enhancement is for a new extract that leverages the existing report's SQL. The new report:
Should contain time entry information for every active employee in the company. (The existing Time Entry report only shows one employee's information at a time. We want all employees to be part of this new job)
Add a new column after "Total Hours" entitled "Hourly Rate". Pull the project's Hourly Rate for the individual if the "Billing" Type is "Hourly". If the "Billing" Type is "Fixed" then the "Hourly Rate" value should be NULL. If the project code is an internal project code, then the "Hourly Rate" value should be NULL. The "Hourly Rate" column should be formatted as currency (e.g $185.00).

Schedule the query to run automatically at 7:00 a.m. on the third day after the close of the current pay period (i.e., on the 3rd and 18th of every month). The report should be automatically emailedl to Finance@logic2020.com without needing user intervention. The email should be entitled "Payroll Distribution Report for <Previous Payroll Period Start Date in MM/DD/YYYY format> - <Previous Payroll Period End Date in MM/DD/YYYY format>"