Comments - Defect Number: 3110; Defect RecordID: 3106; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Projects - Summary: Need to use Actual Revenue and Actual Margin when calculating values for prior months; Defect Description: In the Project Summary page, currently the Revenue and Margin calculations use values from the Project Details page to forecast Revenue and Margin values. This works fine if a user is looking at the current month and future months to see Project Summary financial forecasts. However, with our robust Time Entry we ought to be leveraging the actual time entry data to drive our Revenue and Margin calculations for activities in the past. Therefore, I am requesting an enhancement to the Project Summary page and calculations to incorporate our Time Entry data to calculate Actual Revenue and Actual Margin any time we view the Project Summary results for past months.
To increase flexibility, I'd like to add a checkbox under the "Show Projects for ..." entry called, "Use Actual Time Entry data for Previous Month calculations" that will enable a user to view results using only Projected Values (i.e. to produce the report as it runs today) if turned off and will use Time Entry data for previous month calculations if turned on. The default value of this checkbox will be "On".

When the report runs using the Time Entry data to calculate Revenue and Margin, a previous month is defined as any month preceding the current month (i.e. the current month will always use the Projected Values regardless of whether the checkbox is on or off and regardless of whether Time Entry data exists for the Current Month). This will allow a settling period for the current month's time entry.

For Projects with Fixed Fee pricing, the fixed fee Revenue would be applied to a month's revenue calculation based on the percent of completion of the Milestone. (E.g. Milestone Start Date = 10/1/2012 and End Date is 11/15/2012 and is a fixed price milestone with a payment = $12,000. If a user runs the Project Summary in December 2012, November's Revenue will include $4,000 in fixed fee revenue while October's Revenue reflects $8,000 in Revenue. NOTE the above is simplified for explanatory purposes only, assuming an equal distribution of hours worked over the six week project.) I believe this is how PM functions now for calculating Revenue of Fixed Fee milestones on the Project Summary report. Please correct me if I am wrong.

The Export of Project Summary should also be adjusted. See the attached file as an example new file. The Export will be the same file regardless of whether the "Use Actual Time Entry data for Previous Month calculations" checkbox is On/Off as the Export will include both Projected Values for Revenue and Cont. Margin and Actual Values. The Export will contain two worksheets - one for Summary - Projected and one for Summary - Actuals.

The Project Summary page's results should be pre-cached each night to maximize performance when viewed each day. The caching should pre-build the report views for data within the Current Fiscal Year and the Previous Fiscal Year. This should expedite performance of the displayed data on the screen and limit errors when running the report across large time periods.