Comments - Defect Number: 3115; Defect RecordID: 3111; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Compensation Record - Unable to edit any Compensation record for an employee; Defect Description: With the recent Add Person Workflow we made an enhancement to prevent a user from editing prior Compensation records - only current compensation record can be edited/deleted. I found a use case we didn't previously encounter which when it occurs causes all records to be uneditable.
Steps to Reproduce:
After identifying a consultant, on the Persons tab, change a person's Title. (e.g. from Consultant to Senior Consultant).
This generates a new compensation record with the start date = today's date.
The prior compensation record has an end date = [today's date - 1] and is un-editable.
The new compensation record is editable and can be deleted.
Delete the new compensation record.
Notice that all compensation records are now uneditable as the system did not reactivate the prior compensation record - it left it with an end date = [today's date - 1]. Previously, when a user deleted the most current compensation record, the system would remove the end date from the prior record making it the active record again.

Expected result: When deleting the most current record the system should remove the end date of the prior record making it active again and thus, enabling it to be edited.

I need the above use case addressed. Additionally, I experienced this with Nick Maddock in production. As a result, he does not have an active compensation record. My desire is that the compensation record that has a start date of 1/1/2013 is active so that it can be edited. I recognize I could simply add a new compensation record with a start date = 1/2/2013 but from a data perspective I don't want that history to be reflected in his Compensation record.