Comments - Defect Number: 3123; Defect RecordID: 3119; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Capabilities - need the ability to mark Active/InActive; Defect Description: On the Capabilities page under Configuration - Company Resources - Capabilities, we need the ability to turn on/off Capabilities similar to the way we do for Practice Areas. We need an isActive type indicator to handle cases where our Capabilities change.
When Active the Capability will appear in the Capabilities picklist on the Project Details page. When Inactive, the Capabiliity will not appear in the Capabilities picklist on the Project Details page.

If a Project has a Capability already selected and saved and an Admin marks that capability Inactive, what should happen?
In this case, I would expect the Capability to show in the Picklist when the project is viewed. If a user deselects the capability from the picklist and clicks Save, they would no longer see the InActive Capability in the picklist the next time they open the picklist. While the user is on the Project Details page, the Inactive Capability would be visible in the picklist. It will only be removed and unselectable once a user clicks Save. The user could always Cancel without saving and the old value would be retained.