Comments - Defect Number: 3136; Defect RecordID: 3132; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Opportunity AND Project Details Pages > Creation of "Business Group, "Pricing List, " and "New/Extension" fields; Defect Description: In order to facilitate more discrete reporting at the Client Account level, we need to add three new fields to both the Opportunity AND Project Details pages in Practice Management. These new fields are for reporting purposes only; they drive no new or automated functionality. All three fields should be added as new columns to all "Project Summary" and "Opportunity Summary" report exports. Please review the requirements below and provide an estimate of level of effort to complete.
1. "Business Group"

The first new field to add to the Opportunity and Project Details pages is entitled "Business Group". The field is a display-only field that displays the value based on appropriate mapping of "Business Unit" to "Business Group" as displayed on the "Business Unit" tab on the "Account Details" page in PM (Configuration > Company Resources > Accounts).That mapping is determined on the current "Business Units" tab on the Account Details page. On the "Business Units" tab, a new column needs to be added entitled "Business Group". When a user edits a line item on the "Business Units" tab, they should be able to select the Business Group from a single-select picklist.

The source of the items in the pick list is a new table called "Business Groups" that will reside on a new tab called "Business Groups" on the Account Details page (i.e., in addition to the current "Projects" and "Business Units" tabs).The table will use the same functionality as the current "Business Units" tab (i.e., should have the same addition, editing, deletion and "Active/Non-Active" checkbox functionality). Each current account in Practice Management except Microsoft should have one default entry in the "Business Groups" table, entitled "Default." "Default" can be edited but cannot be deleted from the list (i.e., the table must have at least one item in it). The exception to the "Default" line item is the Microsoft account, which should have the following values entered into the "Business Groups" table:

Business Groups Table Values for the Microsoft Account:
Products and Corp
SMSG - Shared Services
SMSG - Subs & Strategy


2. "Pricing List"

The second new field to add to the Opportunity and Project Details pages is entitled "Pricing List". The field is a single-select pick list. The source of the pick list is a new table on the Account Details page located on a new tab entitled "Pricing List". That table should have the same type of edit functionality as the "Business Group" table described above.

Each active account in Practice Management except Microsoft should have one default entry in the table entitled "Default." Microsoft is again the exception, and should have the following values listed in the table:

Pricing List Table Values for the Microsoft Account:
IT Consulting
Non-Technical Project Management
Strategy
Application Development

The field is NOT a required field to successfully update the Project Details page. The value selected in "Opportunity Details" should transfer to "Project Details" when the Opportunity is linked to a Project.


3. "New Business/Extension"

The third new field to add to the Opportunity and Project Details pages is entitled "New Business/Extension". This field is a single select pick-list with two options to select from: "New Business" or "Extension". There is no default value, and it is not a required field to complete to save. The value selected in "Opportunity Details" should transfer to "Project Details" when the Opportunity is changed to a Project.