Comments - Defect Number: 3137; Defect RecordID: 3133; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Project Details > New "CSAT" tab; Defect Description: In order to capture and track customer satisfaction ("CSAT") on our project engagements, we are requesting that a new tab be created on the Project Details page. The tab, using the similar functionality already existing in Practice Management, consists of a table where summary information is listed. Click "Add CSAT" opens a screen to enter required data and free text fields.
Business requirements and a mock-up of the new tab are attached. In addition, we are also working on requirements for a reporting component for CSAT, and will provide the specs as soon as we have them drafted.

Please review the mock-up attached and provide us a level of effort estimate to complete this first phase of enhancements.