Comments - Defect Number: 3150; Defect RecordID: 3146; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: SQL query needed - pull all W2 Salaried employees and the projects they worked on from 1/1/2013 through 6/1/2013; Defect Description: I am almost positive we've requested this in the past but I cannot locate an e-mail or previous TTP ticket on this request:
To support our Mid-Year Performance Review process I am looking to capture the W2-Salaried employees in the Consulting Division and what project(s) they worked on between the dates Jan. 1, 2013 and June 1, 2013. I'd like the data in an Excel export with the following details:

Pay Type (W2 Salary only)
Division (Consulting only)
Hire Date
Person ID
First Name
Last Name
Account
Business Group
Business Unit
Project # (e.g. P012345)
Project Owner
Start Date (based on the milestone when did they start on the project)
End Date (based on the milestone when did they end on the project)

Many people will have multiple rows each as they will have worked on a variety of projects and potentially accounts during the 1/1/2013 to 6/1/2013 period.

For the Start Date and End Date values, some resources may have strange schedules. Therefore, it would be sufficient to simply put the earliest date they were assigned to the project and the latest date. For example, if someone was assigned to Project X as of 1/5/2013 through 1/21/2013 and then again from 2/25/2013 to 2/28/2013. I would be okay with seeing 1/5/2013 in the Start Date column and 2/28/2013 in the End Date column.

This would probably make sense to setup as a Stored Procedure as we will need this again next year if a procedure does not already exist.

I hope this is clear but if not do not hesitate to ask questions. I'd love to have this early in the week next week (May 28th) if at all possible.

Thanks,
Matt