Comments - Defect Number: 3147; Defect RecordID: 3143; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Accounts > "House Account" designation; Defect Description: As mentioned before, Logic20/20 has kicked off an initiative to improve our financial reporting. We successfully launched the first version of an Excel-based model that tracks and reports on company financials across a variety of pivots (the Attainment report export you created for us last week is the main data source that drives this model). Beyond tracking the company's progress against revenue targets and tracking attainment of those targets, the model will be used to assign credit (and therefore compensation) for Salespersons and Client Directors based on the number of projects assigned to them (using the Project Details page as the data input).
There some accounts which are not owned by a specific Salesperson or Client Director and thus we don't want those accounts counted toward the Salesperson's and Client Director's revenue attainment. As a result, we wish to add functionality into PM that will designate some client accounts as "House Accounts" in order to designate this distinction.

For each Account on the "Account Details" page, please add a checkbox under "Account Active?" entitled "House Account?".

The addition of the House Account drives the following changes to the Project Summary and the Attainment Report exports:

On the Project Summary Export report (both "Summary - Projected" and "Summary - Active" tabs), add a new column between "Account" and "Business Group" entitled "House Account". If the House Account checkbox on the Account Details page is checked, then enter "Yes" for every project that is a belongs to the selected Account. If the checkbox has not been selected, then nulll.
On the Attainment Export, if the House Account checkbox on the Account Details page is checked, then replace the Salesperson's and the Director's name with "House Account"

Please let me know if you have any questions!