Comments - Defect Number: 3162; Defect RecordID: 3158; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Time Reports error: Inconsistent Available Hours results on By Time Period and By Person report views; Defect Description: The Available Hours displayed on the By Time Period and the By Person report views are inconsistent when the Q2 time range is selected.
Steps to Reproduce in Production:
1. Run the By Time Period report - selecting "Q2" as the date range and select View: By Resource
2. Filter results to show Sr. Managers
3. Locate the result for "Arry Yu". For Arry, her total billable hours is 230 and available hours is 520. Billable utilization is reflected as 54.2%, but a manual calculation of billable hours/available hours as per this view (230/520) equals 44.2%.
4. Click the hyperlink under 54.2%. The By Person results report will display. Hovering over the Billable Utilization percentage at the top of the screen displays the utilization calculation, reflecting 424 hours as the available hours in the time period.

Expected result: Available hours are calculated within the time period as the start of the time period to the current date. For example, Q2 available hours as of today, 6/13/2013, would be calculated as from 4/1/2013 - 6/13/2013, not until the end of the month on 6/30/2013. This is universally true for all Available Hours calculations for all time reports and related extracts. I tested specifically the Q2 time range, and do not know if this error is impacting other time ranges as well. I suspect the error is a product of how to calculate Available Hours for a time range that has not yet been completed, but would appreciate your investigation.

Observed result: It appears that the By Person time report is calculating Available Hours correctly, and thus the billable utilization calculation appears not to be impacted.

Could you please investigate this issue and push the fix to Production.